7 reactive buy sell dialogs (#10)

+++ domain/src/main/java/ru/surf/learn2invest/domain/domain_models/Transaction.kt
@@ -22,6 +22,6 @@ data class Transaction(
     val symbol: String,
     val coinPrice: Float,
     val dealPrice: Float,
-    val amount: Float,
+    val amount: Int,
     val transactionType: TransactionsType
 )
